Enable JitTrace test in release branch
