+++ test/WebJobs.Script.Tests.Integration/WebHostEndToEnd/SpecializationE2ETests.cs
@@ -443,25 +443,6 @@ namespace Microsoft.Azure.WebJobs.Script.Tests
             }
         }
 
-        [Theory(Skip = "Will be enforced as part of release build")]
-        [InlineData(WarmUpConstants.JitTraceFileName)]
-        [InlineData(WarmUpConstants.LinuxJitTraceFileName)]
-        public void ColdStart_JitFailuresTest(string fileName)
-        {
-            var path = Path.Combine(Path.GetDirectoryName(new Uri(typeof(HostWarmupMiddleware).Assembly.CodeBase).LocalPath), WarmUpConstants.PreJitFolderName, fileName);
-
-            var file = new FileInfo(path);
-
-            Assert.True(file.Exists, $"Expected PGO file '{file.FullName}' does not exist. The file was either renamed or deleted.");
-
-            JitTraceRuntime.Prepare(file, out int successfulPrepares, out int failedPrepares);
-
-            var failurePercentage = (double) failedPrepares / successfulPrepares * 100;
-            
-            // using 1% as approximate number of allowed failures before we need to regenrate a new PGO file.
-            Assert.True( failurePercentage < 1.0 , $"Number of failed PGOs are more than 1 percent! Current number of failures are {failedPrepares}. This will definitely impact cold start! Time to regenrate PGOs and update the {fileName} file!");
-        }
-
         private IWebHostBuilder CreateStandbyHostBuilder(params string[] functions)
         {
             string scriptRootConfigPath = ConfigurationPath.Combine(ConfigurationSectionNames.WebHost, nameof(ScriptApplicationHostOptions.ScriptPath));
